tools/mtp-extract/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mtp-extract"
version = "0.7.0"
description = "Extraction CLI for generating draft MTP packages from conversations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Lubor Fedak", email = "lubor@codemoravia.com" },
]
keywords = ["mtp", "extraction", "conversation", "provenance", "methodology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "mtp-lint",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
mtp-extract = "mtp_extract.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mtp_extract"]
